doc-update-assistant/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
orjson>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
python-multipart>=0.0.6
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
numpy>=1.25.0